Rag  Platform/rag-platform/rag-platform/requirements.txt
# Core RAG stack
langchain==0.3.14
langchain-community==0.3.14
langchain-chroma==0.2.0
langchain-openai==0.3.0
chromadb==0.6.3

# Embeddings (local, no API key needed)
sentence-transformers==3.3.1

# Document loaders
pypdf==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0

# LLM backends (install what you need)
# ollama — install the desktop app from https://ollama.com, then: ollama pull mistral
# For HuggingFace Hub inference:
# huggingface-hub==0.27.0
# For local HuggingFace Transformers:
# transformers==4.48.0
# accelerate==1.3.0
# bitsandbytes==0.45.0  # 4-bit quantization (Linux/WSL only on Windows)

# API server
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# Utilities
python-dotenv==1.0.1
